+++ CHANGELOG.md
@@ -6,6 +6,13 @@ FEATURES:
 - `ksyun_vpc` 支持IPv6网段
 - `ksyun_subnet` 支持IPv6网段
 
+## 1.5.0 (Jul 12, 2023)
+FEATURES:
+
+- - **New Resource:** `ksyun_knad` knad 原生高防
+- - **New Resource:** `ksyun_knad_associate`
+- - **New Data Source:** `ksyun_knads` 
+
 ## 1.4.2 (Jul 07, 2023)
 
 IMPROVEMENTS: